+++ kipper/cli/src/commands/analyse.ts
@@ -4,10 +4,10 @@
  */
 import { Command, flags } from "@oclif/command";
 import { KipperCompiler, KipperError, KipperLogger, KipperParseStream, LogLevel } from "@kipper/core";
-import { KipperEncoding, KipperEncodings, KipperParseFile, verifyEncoding } from "../file-stream";
+import { KipperEncodings, KipperParseFile, verifyEncoding } from "../file-stream";
 import { CLIEmitHandler, defaultCliLogger } from "../logger";
-import { KipperInvalidInputError } from "../errors";
 import { IFlag } from "@oclif/command/lib/flags";
+import { getFile } from "../compile";
 
 export default class Analyse extends Command {
 	static override description = "Analyse a Kipper file and validate its syntax and semantic integrity.";
@@ -48,28 +48,21 @@ export default class Analyse extends Command {
 		const compiler = new KipperCompiler(logger);
 
 		// Fetch the file
-		let file: KipperParseFile | KipperParseStream;
-		if (args.file) {
-			file = await KipperParseFile.fromFile(args.file, flags["encoding"] as KipperEncoding);
-		} else if (flags["string-code"]) {
-			file = await new KipperParseStream(flags["string-code"]);
-		} else {
-			throw new KipperInvalidInputError("Argument 'file' or flag '-s/--string-code' must be populated. Aborting...");
-		}
+		let file: KipperParseFile | KipperParseStream = await getFile(args, flags);
+
+		// Compilation configuration
+		const parseStream = new KipperParseStream({
+			name: file.name,
+			filePath: file instanceof KipperParseFile ? file.absolutePath : file.filePath,
+			charStream: file.charStream,
+		});
 
 		// Start timer for processing
 		const startTime: number = new Date().getTime();
 
 		// Analyse the file
 		try {
-			await compiler.syntaxAnalyse(
-				new KipperParseStream(
-					file.stringContent,
-					file.name,
-					file instanceof KipperParseFile ? file.absolutePath : file.filePath,
-					file.charStream,
-				),
-			);
+			await compiler.syntaxAnalyse(parseStream);
 
 			// Finished!
 			const duration: number = (new Date().getTime() - startTime) / 1000;